Add files via upload
In the first workshop I have to read in 7 variables(Str,Dex,Con,Int,Wis,Cha and Level) from the console. Basically I have to prompt these variables from the user. After read in the variables then the program should calculate the bonus as follows. Bonus should be 0 at 10, cumulative +1 for each even number above 10 and -1 for each odd number below 10. And then Hit Points should be calculated using following equation. Hit Points = level * level * (Math.random () * 1000 % 6 + 1). If the bonus is a positive number display + before the number, if the number is a negative number display – before the number and no sign if the bonus is 0.
If the level is less than or equal 0 display an error message and ask the user to re-enter the level value. And also if the other variables values are less than 0 displays an error message and ask the user to re-enter the specific value.
At last I have to display the output